Simulation Report addendum
Given that the entire reason I settled on there being 4 entries in this codex was because of how it originally was crafted with 4 data fragments, I felt that this would be a good opportunity to add a bonus entry to account for needing 5 data fragments to craft now. Admittedly this addendum does imply that the Ancients are a successful group of Glitch/Simulacrum (which also means that the Deadbeat predate the Ancients), so you're free to disregard this if it doesn't quite work for you.